Rename part 2
Moved header files from httpmicroservice folder to skye.
