+++ examples/echo.cpp
@@ -1,4 +1,4 @@
-#include <httpmicroservice/service.hpp>
+#include <skye/service.hpp>
 
 #include <boost/algorithm/string/case_conv.hpp>
 #include <boost/asio/signal_set.hpp>
