+++ examples/hello.cpp
@@ -1,4 +1,4 @@
-#include <httpmicroservice/service.hpp>
+#include <skye/service.hpp>
 
 namespace asio = boost::asio;
 namespace http = boost::beast::http;
