+++ examples/producer.cpp
@@ -1,5 +1,5 @@
-#include <httpmicroservice/format.hpp>
-#include <httpmicroservice/service.hpp>
+#include <skye/format.hpp>
+#include <skye/service.hpp>
 
 #include <boost/asio/signal_set.hpp>
 #include <boost/asio/thread_pool.hpp>
