--- include/httpmicroservice.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <httpmicroservice/format.hpp>
-#include <httpmicroservice/service.hpp>
-#include <httpmicroservice/session.hpp>
-#include <httpmicroservice/types.hpp>
