+++ tests/test.cpp
@@ -1,4 +1,4 @@
-#include <httpmicroservice.hpp>
+#include <skye.hpp>
 
 #include <catch2/catch_test_macros.hpp>
