+++ tests/test_service.cpp
@@ -1,4 +1,4 @@
-#include <httpmicroservice/service.hpp>
+#include <skye/service.hpp>
 
 #include <catch2/catch_test_macros.hpp>
