+++ tests/test_session.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include <httpmicroservice/session.hpp>
+#include <skye/session.hpp>
 
 #include "mock_sock.hpp"
 #include "test.hpp"
